Create template and make hierarchy.
